+++ Poker Chip/Player.swift
@@ -11,18 +11,18 @@ import SwiftData
 struct Player: Identifiable, Codable, Hashable {
     var id: UUID = UUID()
     var name: String
-    var chip: Double
+    var chip: Decimal
     var playerRecord: PlayerRecord?
     var position: String = ""
     var fold: Bool = false
-    var raiseSize: Double = 0.0
-    var potLimit: Double = 0.0
+    var raiseSize: Decimal = 0.0
+    var potLimit: Decimal = 0.0
     var actionStr: String = ""
     var listIndex: Int = 0
-    var buyIn: Double = 0.0
+    var buyIn: Decimal = 0.0
     var curPlayerAnimation: Bool = false
     var VPIPCurRound: Bool = false
     var PFRCurRound: Bool = false
-    var curRoundWinningReset: Double = 0.0
-    var curRoundPlayerRecordWinningReset: Double = 0.0
+    var curRoundWinningReset: Decimal = 0.0
+    var curRoundPlayerRecordWinningReset: Decimal = 0.0
 }
