+++ Poker Chip/PlayerList.swift
@@ -9,8 +9,8 @@ import Foundation
 
 struct PlayerList: Codable {
     var playerList: [Player] = [Player]()
-    var pot: Double = 0.0
-    var blinds: [Double] = [Double]()
+    var pot: Decimal = 0.0
+    var blinds: [Decimal] = [Decimal]()
     func data() -> Data? {
         try? JSONEncoder().encode(self)
     }
